chore: general update/addition of files at project root level (#427)
The changes include:
- an update of the content of some files at project root level, e.g.
AUTHORS, CONTRIBUTING, ...
- remove black, isort and flake8 to use ruff instead
- add .gitattributes to ease working from different OS environments
- add/update tools configuration, e.g. codecov (which will be leveraged
later), pre-commit, ruff

I also removed Python 3.8 from tox list of targets.

Is there a reason not to state that PySpeos is compatible with python
3.12 and 3.13 ? If not, I'll check if it's already working with this
versions or try to make it work and update the project accordingly.

---

> [!NOTE]  
The current configuration of `ruff` is ignoring a lot of rules that
shouldn't be ignored. This is a task that should be handled by multiple
people as it is really time consuming. I started commented and fixing
some of them but stopped on the way.

---------

Co-authored-by: sthoene <stefan.thoene@ansys.com>
Co-authored-by: Revathy Venugopal <104772255+Revathyvenugopal162@users.noreply.github.com>
Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>