docs: add example assets button (#518)
## Description
Adding a button to download the examples assets. I had too add a work
around to ensure that the 3 download buttons were aligned somehow
"correctly". See the result below


![image](https://github.com/user-attachments/assets/58fcdb8f-21a0-4416-9b45-de895c5d4a35)

If that's fine with you, we can leave it like that for the moment as we
have to test that assets are correctly uploaded and available for
download through the documentation. A next step (PR) could be to remove
the buttons and only use text content with hyperlinks to trigger the
downloading of files.

## Issue linked
Related to #497 
Close #517 

## Checklist
- [x] I have tested my changes locally.
- [x] I have added necessary documentation or updated existing
documentation.
- [x] I have followed the coding style guidelines of this project.
- [ ] I have added appropriate tests (unit, integration, system).
- [x] I have reviewed my changes before submitting this pull request.
- [x] I have linked the issue or issues that are solved by the PR if
any.
- [x] I have assigned this PR to myself.
- [x] I have made sure that the title of my PR follows [Conventional
commits style](https://www.conventionalcommits.org/en/v1.0.0/#summary)
(e.g. ``feat: add optical property``)
- [x] I have agreed with the Contributor License Agreement
([CLA](https://developer.ansys.com/form/cla-acceptance)).

---------

Co-authored-by: pyansys-ci-bot <92810346+pyansys-ci-bot@users.noreply.github.com>
Co-authored-by: sthoene <stefan.thoene@ansys.com>